ISSUE #4780 - move revisions to containers/drawings
