+++ frontend/src/v4/modules/sagas.ts
@@ -23,13 +23,13 @@ import teamspaces2Saga from '@/v5/store/teamspaces/teamspaces.sagas';
 import currentUser2Saga from '@/v5/store/currentUser/currentUser.sagas';
 import auth2Saga from '@/v5/store/auth/auth.sagas';
 import usersSaga from '@/v5/store/users/users.sagas';
-import containerRevisionsSaga from '@/v5/store/containerRevisions/containerRevisions.sagas';
+import containerRevisionsSaga from '@/v5/store/containers/containerRevisions/containerRevisions.sagas';
 import ticketsSaga from '@/v5/store/tickets/tickets.sagas';
 import ticketCommentsSaga from '@/v5/store/tickets/comments/ticketComments.sagas';
 import viewer2Saga from '@/v5/store/viewer/viewer.sagas';
 import ticketsCardSaga from '@/v5/store/tickets/card/ticketsCard.sagas';
 import drawingsSaga from '@/v5/store/drawings/drawings.sagas';
-import drawingRevisionsSaga from '@/v5/store/drawingRevisions/drawingRevisions.sagas';
+import drawingRevisionsSaga from '@/v5/store/drawings/drawingRevisions/drawingRevisions.sagas';
 import activitiesSaga from './activities/activities.sagas';
 import authSaga from './auth/auth.sagas';
 import billingSaga from './billing/billing.sagas';
