+++ frontend/src/v5/services/selectorsHooks.ts
@@ -20,10 +20,10 @@ import * as AuthSelectors from '@/v5/store/auth/auth.selectors';
 import * as BimSelectors from '@/v4/modules/bim/bim.selectors';
 import * as CanvasHistorySelectors from '@/v4/modules/canvasHistory/canvasHistory.selectors';
 import * as ContainersSelectors from '@/v5/store/containers/containers.selectors';
-import * as ContainerRevisionsSelectors from '@/v5/store/containerRevisions/containerRevisions.selectors';
+import * as ContainerRevisionsSelectors from '@/v5/store/containers/containerRevisions/containerRevisions.selectors';
 import * as CurrentUserSelectors from '@/v5/store/currentUser/currentUser.selectors';
 import * as DrawingsSelectors from '@/v5/store/drawings/drawings.selectors';
-import * as DrawingRevisionsSelectors from '@/v5/store/drawingRevisions/drawingRevisions.selectors';
+import * as DrawingRevisionsSelectors from '@/v5/store/drawings/drawingRevisions/drawingRevisions.selectors';
 import * as ProjectsSelectors from '@/v5/store/projects/projects.selectors';
 import * as FederationsSelectors from '@/v5/store/federations/federations.selectors';
 import * as GroupsSelectors from '@/v4/modules/groups/groups.selectors';
