+++ frontend/src/v5/store/reducers.ts
@@ -18,7 +18,7 @@
 import { federationsReducer } from '@/v5/store/federations/federations.redux';
 import { containersReducer } from '@/v5/store/containers/containers.redux';
 import { drawingsReducer } from '@/v5/store/drawings/drawings.redux';
-import { drawingRevisionsReducer } from '@/v5/store/drawingRevisions/drawingRevisions.redux';
+import { drawingRevisionsReducer } from '@/v5/store/drawings/drawingRevisions/drawingRevisions.redux';
 import { dialogsReducer } from '@/v5/store/dialogs/dialogs.redux';
 import { projectsReducer } from '@/v5/store/projects/projects.redux';
 import { teamspacesReducer } from '@/v5/store/teamspaces/teamspaces.redux';
@@ -26,7 +26,7 @@ import { usersReducer } from '@/v5/store/users/users.redux';
 import { currentUserReducer } from '@/v5/store/currentUser/currentUser.redux';
 import { authReducer } from '@/v5/store/auth/auth.redux';
 import { ticketsReducer } from '@/v5/store/tickets/tickets.redux';
-import { containerRevisionsReducer } from '@/v5/store/containerRevisions/containerRevisions.redux';
+import { containerRevisionsReducer } from '@/v5/store/containers/containerRevisions/containerRevisions.redux';
 import { ticketsCardReducer } from '@/v5/store/tickets/card/ticketsCard.redux';
 import { ticketCommentsReducer } from '@/v5/store/tickets/comments/ticketComments.redux';
 import { viewerReducer } from '@/v5/store/viewer/viewer.redux';
