+++ frontend/src/v5/ui/routes/dashboard/projects/containers/uploadContainerRevisionForm/sidebarForm/sidebarForm.component.tsx
@@ -23,7 +23,7 @@ import * as countriesAndTimezones from 'countries-and-timezones';
 import { MenuItem } from '@mui/material';
 import { FormSelect, FormTextField } from '@controls/inputs/formInputs.component';
 import { get } from 'lodash';
-import { LOD_VALUES } from '@/v5/store/containerRevisions/containerRevisions.types';
+import { LOD_VALUES } from '@/v5/store/containers/containerRevisions/containerRevisions.types';
 import { Heading, AnimationsCheckbox, TimezoneSelect, Title, FlexContainer, HiddenMenuItem } from './sidebarForm.styles';
 import { extensionIsRevit, extensionIsSpm } from '../extensions.helpers';
 import { useContext } from 'react';
