+++ frontend/src/v5/validation/drawingSchemes/validators.ts
@@ -17,7 +17,7 @@
 import * as Yup from 'yup';
 import { formatMessage } from '@/v5/services/intl';
 import { alphaNumericHyphens } from '../shared/validators';
-import { selectRevisions } from '@/v5/store/drawingRevisions/drawingRevisions.selectors';
+import { selectRevisions } from '@/v5/store/drawings/drawingRevisions/drawingRevisions.selectors';
 import { getState } from '@/v4/modules/store';
 
 export const revisionName = Yup.string()
